+++ deltafi-action-kit/src/test/java/org/deltafi/actionkit/action/egress/EgressResultTest.java
@@ -1,0 +1,60 @@
+/*
+ *    DeltaFi - Data transformation and enrichment platform
+ *
+ *    Copyright 2022 DeltaFi Contributors <deltafi@deltafi.org>
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+
+package org.deltafi.actionkit.action.egress;
+
+import org.deltafi.common.types.ActionContext;
+import org.deltafi.common.types.Metric;
+import org.junit.jupiter.api.Test;
+
+import static org.hamcrest.MatcherAssert.*;
+import static org.hamcrest.Matchers.*;
+
+import java.util.List;
+
+public class EgressResultTest {
+
+    final String DESTINATION = "overThere";
+    final long BYTES_EGRESSED = 42;
+    ActionContext actionContext = new ActionContext("myDid", "myName", "myIngressFlow", "myEgressFlow", "myHostName", "myActionVersion");
+    EgressResult sut = new EgressResult(actionContext, DESTINATION, BYTES_EGRESSED);
+    
+    @Test
+    void testDefaultCustomMetrics() {
+        List<Metric> metrics = sut.getCustomMetrics();
+
+        assertThat(metrics, containsInAnyOrder(
+                Metric.builder().name("files_out").value(1).build().addTag("endpoint", DESTINATION),
+                Metric.builder().name("bytes_out").value(BYTES_EGRESSED).build().addTag("endpoint", DESTINATION)
+        ));
+    }
+
+    @Test
+    void testAdditionalCustomMetrics() {
+        Metric additionalMetric = new Metric("wig_out", 9).addTag("starlord", "was here");
+        sut.add(additionalMetric);
+        List<Metric> metrics = sut.getCustomMetrics();
+
+        assertThat(metrics, containsInAnyOrder(
+                Metric.builder().name("files_out").value(1).build().addTag("endpoint", DESTINATION),
+                Metric.builder().name("bytes_out").value(BYTES_EGRESSED).build().addTag("endpoint", DESTINATION),
+                additionalMetric
+        ));
+    }
+
+}